requirements.txt
Django
djangorestframework
django-cors-headers
requests
python-dotenv
groq